+++ src/extensionFunctions.ts
@@ -49,6 +49,7 @@ export const GATHER_ALL_COMMANDS = 'vscode.didact.gatherAllCommands';
 export const VIEW_OPEN_TUTORIAL_MENU = 'vscode.didact.view.tutorial.open';
 export const REGISTER_TUTORIAL = 'vscode.didact.register'; // name, uri, category
 export const REFRESH_DIDACT_VIEW = 'vscode.didact.view.refresh';
+export const SEND_TERMINAL_KEY_SEQUENCE = 'vscode.didact.sendNamedTerminalCtrlC';
 
 export const DIDACT_OUTPUT_CHANNEL = 'Didact Activity';
 
@@ -161,7 +162,15 @@ export namespace extensionFunctions {
 		}
 	}
 
-	function findTerminal(name: string) : vscode.Terminal | undefined {
+	async function showAndSendCtrlC(terminal: vscode.Terminal) {
+		if (terminal) {
+			terminal.show();
+			await vscode.commands.executeCommand("workbench.action.terminal.sendSequence", { text : "\x03" });
+			return;
+		}
+	}
+
+	export function findTerminal(name: string) : vscode.Terminal | undefined {
 		for(let localTerm of vscode.window.terminals){
 			if(localTerm.name === name){ 
 				return localTerm; 
@@ -183,6 +192,16 @@ export namespace extensionFunctions {
 		sendTextToOutputChannel(`Sent terminal ${name} the text ${text}`);
 	}
 
+	export async function sendTerminalCtrlC(name:string) {
+		const terminal : vscode.Terminal | undefined = findTerminal(name);
+		if (!terminal) {
+			sendTextToOutputChannel(`No terminal found with name ${name} to send a Ctrl+C`);
+		} else {
+			showAndSendCtrlC(terminal);
+			sendTextToOutputChannel(`Sent terminal ${name} a Ctrl+C`);
+		}
+	}
+	
 	// reset the didact window to use the default set in the settings
 	export async function openDidactWithDefault() {
 		sendTextToOutputChannel(`Starting Didact window with default`);