+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "vscode-didact" extension will be documented in this 
 ## 0.1.6
 
 - Added new preference to disable the default information notifications when didact commands are invoked. Links specifying explicit completion messages are still displayed, but links without are not. Preference is set to true by default.
+- Added a sendNamedTerminalCtrlC command to enable stopping a long-running command in a terminal window.
 
 ## 0.1.5